+++ src/retriable/settings.cr
@@ -2,9 +2,9 @@ require "habitat"
 
 module Retriable
   Habitat.create do
-    setting max_attempts : Int32?
+    setting max_attempts : UInt64?
     setting on : Exception.class | Array(Exception.class) | Nil
-    setting on_retry : Proc(Exception, Int32, Time::Span, Time::Span, Nil)?
+    setting on_retry : Proc(Exception, UInt64, Time::Span, Time::Span, Nil)?
     setting base_interval : Time::Span = 0.5.seconds
     setting max_elapsed_time : Time::Span = 15.minutes
     setting max_interval : Time::Span = 1.minute
@@ -17,14 +17,14 @@ module Retriable
   end
 
   class Settings
-    def self.on_retry=(block : (Exception, Int32, Time::Span, Time::Span) -> _)
-      @@on_retry = ->(ex : Exception, attempt : Int32, elapsed_time : Time::Span, interval : Time::Span) do
+    def self.on_retry=(block : (Exception, UInt64, Time::Span, Time::Span) -> _)
+      @@on_retry = ->(ex : Exception, attempt : UInt64, elapsed_time : Time::Span, interval : Time::Span) do
         block.call(ex, attempt, elapsed_time, interval)
         nil
       end
     end
 
-    def self.on_retry(&block : (Exception, Int32, Time::Span, Time::Span) -> _)
+    def self.on_retry(&block : (Exception, UInt64, Time::Span, Time::Span) -> _)
       self.on_retry = block
     end
   end
